Improve uvm-generate-modules-json command
Description
===========

Move away from docopts for development/testing command
`uvm-generate-modules-json`.
