+++ .gitignore
@@ -1,6 +1,6 @@
 # Generated by Cargo
 # will have compiled files and executables
-/target/
+target/
 .cargo/
 
 # Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
