+++ commands/uvm-generate-modules-json/Cargo.toml
@@ -2,8 +2,11 @@
 name = "uvm-generate-modules-json"
 version = "1.0.0"
 authors = ["Manfred Endres <manfred.endres@tslarusso.de>"]
+description = "Write the modules.json for the given unity version."
 edition = "2018"
+
 [dependencies]
+anyhow = "1.0.38"
 cfg-if = "1.0.0"
 flexi_logger = "0.16.1"
 indicatif = "0.15.0"
@@ -12,6 +15,7 @@ console = "0.9.1"
 serde = "1.0"
 serde_derive = "1.0"
 serde_json = "1.0.59"
+structopt = "0.3"
 lazy_static = "1.3.0"
 uvm_cli = { path = "../../uvm_cli" }
 uvm_core = { path = "../../uvm_core" }
