+++ commands/uvm-generate-modules-json/src/main.rs
@@ -1,35 +1,125 @@
+use anyhow::Result;
+use std::io::Write;
+use std::fs::File;
+use std::path::PathBuf;
+use uvm_core::Version;
 use console::style;
 use log::{info, trace};
 use std::io;
-use std::process;
 use uvm_cli;
 use uvm_core;
 use uvm_core::unity::Manifest;
-use uvm_generate_modules_json::Options;
+use uvm_cli::{options::ColorOption, set_colors_enabled, set_loglevel};
+use structopt::{clap::AppSettings, clap::crate_authors, clap::crate_description, clap::crate_version, StructOpt};
 
-const USAGE: &str = "
-uvm-generate-modules-json - Write the modules.json for the given unity version.
+const SETTINGS: &'static [AppSettings] = &[AppSettings::ColoredHelp, AppSettings::DontCollapseArgsInUsage];
 
-Usage:
-  uvm-generate-modules-json [options] <version>...
-  uvm-generate-modules-json (-h | --help)
+#[derive(StructOpt, Debug)]
+#[structopt(version = crate_version!(), author = crate_authors!(), about = crate_description!(), settings = SETTINGS)]
+struct Opts {
+    #[structopt(name = "version", min_values(1), required(true))]
+    version: Vec<Version>,
 
-Options:
-  -o=PATH, --output-dir=PATH        the output path. default stdout
-  -n=NAME, --name=NAME              name of the output file. [default: unity-{version}.json]
-  -f, --force                       force override of existing files.
-  -v, --verbose                     print more output
-  -d, --debug                       print debug output
-  --color WHEN                      Coloring: auto, always, never [default: auto]
-  -h, --help                        show this help message and exit
-";
+    /// the output path. default stdout
+    #[structopt(short, long)]
+    output_dir: Option<PathBuf>,
 
-fn generate_modules(options: Options) -> io::Result<()> {
-    for version in options.version() {
-        let mut output_handle = options.output(&version)?;
-        let output_path = options.output_path(&version);
+    /// name of the output file.
+    #[structopt(short, long, default_value = "unity-{version}.json")]
+    name: String,
 
-        let manifest = Manifest::load(version).map_err(|_| {
+    /// force override of existing files.
+    #[structopt(short, long)]
+    force: bool,
+
+    /// print debug output
+    #[structopt(short, long)]
+    debug: bool,
+
+    /// print more output
+    #[structopt(short, long, parse(from_occurrences))]
+    verbose: i32,
+
+    /// Color:.
+    #[structopt(short, long, possible_values = &ColorOption::variants(), case_insensitive = true, default_value)]
+    color: ColorOption,
+}
+
+pub enum Output {
+    File(File),
+    Stdout,
+}
+
+impl Default for Output {
+    fn default() -> Self {
+        Output::Stdout
+    }
+}
+
+impl Write for Output {
+    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
+        use Output::*;
+        match self {
+            File(x) => x.write(buf),
+            _ => console::Term::stdout().write(buf),
+        }
+    }
+
+    fn flush(&mut self) -> io::Result<()> {
+        use Output::*;
+        match self {
+            File(x) => x.flush(),
+            _ => console::Term::stdout().flush(),
+        }
+    }
+}
+
+impl Opts {
+    pub fn output<V: AsRef<Version>>(&self, version:V) -> io::Result<impl Write> {
+        use std::fs::OpenOptions;
+        if let Some(path) = &self.output_dir {
+            if !path.is_dir() {
+                Err(io::Error::new(io::ErrorKind::InvalidInput, "output path is not a directory"))
+            } else {
+                OpenOptions::new()
+                    .write(true)
+                    .create(true)
+                    .create_new(!self.force)
+                    .open(path.join(self.name(version)))
+                    .and_then(|f| Ok(Output::File(f)))
+            }
+        } else {
+            Ok(Output::default())
+        }
+    }
+
+    pub fn name<V: AsRef<Version>>(&self, version:V) -> String {
+        let name = &self.name;
+        let version = version.as_ref();
+        name.as_str().replace("{version}", &version.to_string())
+    }
+
+    pub fn output_path<V: AsRef<Version>>(&self, version:V) -> Option<PathBuf> {
+        let path = self.output_dir.as_ref()?;
+        if !path.is_dir() {
+            None
+        } else {
+            Some(path.join(self.name(version)))
+        }
+    }
+}
+
+fn main() -> Result<()> {
+    let opt = Opts::from_args();
+    set_colors_enabled(&opt.color);
+    set_loglevel(opt.debug.then(|| 2).unwrap_or(opt.verbose));
+    trace!("generate modules.json");
+
+    for version in &opt.version {
+        let mut output_handle = opt.output(&version)?;
+        let output_path = opt.output_path(&version);
+
+        let manifest = Manifest::load(&version).map_err(|_| {
             io::Error::new(io::ErrorKind::NotFound, "Unable to load manifest")
         })?;
 
@@ -38,19 +128,7 @@ fn generate_modules(options: Options) -> io::Result<()> {
         }
         manifest.write_modules_json(&mut output_handle)?;
     }
-    Ok(())
-}
-
-fn main() {
-    let options: Options = uvm_cli::get_options(USAGE).unwrap();
-    trace!("generate modules.json");
-
-    generate_modules(options).unwrap_or_else(|err| {
-        let message = "Unable generate modules.json";
-        eprintln!("{}", style(message).red());
-        eprintln!("{}", style(err).red());
-        process::exit(1);
-    });
-
+    eprintln!("");
     eprintln!("{}", style("Done").green());
+    Ok(())
 }
